requirements.txt
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2
pandas==2.1.1
numpy==1.26.4
urllib3==2.0.5
chardet==5.2.0
idna==3.4
certifi==2023.7.22
boto3==1.28.41
PyJWT==2.8.0
xlsxwriter==3.1.9